dsa: widen the fp8 k-cache quant kernel's token_id to int64
In _quantize_k_cache_fast_kernel, token_id = tl.program_id(0) is int32 by
default, but it scales every pointer offset into the k/scale buffers
(k_nope_ptr + token_id * k_nope_stride_0, and the output/scale pointers).
On a large KV cache where the token index exceeds the int32 range, those
offsets wrap and the kernel reads/writes the wrong memory. Cast to int64
so the pointer arithmetic holds at scale.

+++ python/sglang/kernels/ops/attention/dsa/quant_k_cache.py
@@ -283,7 +283,7 @@ def _quantize_k_cache_fast_kernel(
     FP8_MIN: tl.constexpr,
     FP8_MAX: tl.constexpr,
 ):
-    token_id = tl.program_id(0)
+    token_id = tl.program_id(0).to(tl.int64)
     raw_block_id = tl.program_id(1)
 
     if raw_block_id < NUM_NOPE_BLOCKS:
